Crate game_system

Example crate

Implements the state machine for a board game. Any type which interacts with the built state machine MUST be exported from this crate. Any type that's directly used from medici_core MUST transitively be re-exported here. Re-exporting allows the downstream crates to have a single dependancy, this crate.

See also

[game_rules] for an example how to implement game rules which use the state machine defined within this crate.

Modules

card
Specialized implementation of the Card game-object.

entity
Specialized implementation of EntityStruct for our state machine.

prelude
Often used types exported together for ease of use.

prototype
Specialized prototypes for our state machine.

re_export
Exported types from medici_core.

runtime
Module pending removal

setup
Module containing state machine construction methods.

state_machine
State machine implementation for the game: Goose game.

tag
Module containing the property tags which can be attributed to [Entity]s and other game-objects.

Macros

card_impl
Use this macro to build implementations of new cards.