Module game_system::prelude

Often used types exported together for ease of use.

This module can be imported like so

extern crate game_system;
use game_system::prelude::*;

and will cause all exported types to be injected in the declared scope.

Re-exports

pub use medici_core::ctstack::*;

pub use medici_core::ctxt;

pub use medici_core::hydrate;

pub use card::Card;

pub use entity::Entity;

pub use state_machine::config::SetupConfig;

pub use state_machine::machine::Machine;

pub use state_machine::state::leaf::triggerable::*;

pub use state_machine::state::leaf::*;

pub use state_machine::state::toplevel::*;

pub use tag::EntityTags;

pub use prototype;

pub use state_machine::transaction;

Modules

error
Types, to be used within the system, providing context of unexpected behaviour.

function
Contains the core functionality items for our system.

Structs

CardId
Type that's generally used to identify and order Card objects.

MachineError
User facing error type indicating a failure during evalutation/computation of the state machine.

Enums

ErrorKind
Enumeration of publicl cases of state machine failures.

Constants

GAME_E_ID
The game entity should always have ID 0.

Traits

ArrayStorageCompliance
Defines array access behaviour for storage objects.

CardBuilder
Trait used to create a new Card object.

CardTrait
Trait representing an actual game card.

EntityBuilder
Trait used to create a new Entity object.

EntityTrait
Trait representing an object which properties can be altered dynamically (at runtime).

FrontendErrorExt
Trait facilitating error creation with a snapshot of the state machine attached.

HydratedErrorExt
TODO

Identifiable
Trait enforcing implementing objects to expose their identifier. The uniqueness of this identifier depends on the implementing object itself.

IndexedStorageCompliance
Defines indexed behaviour for a certain storage object.

PullupInto
Syntax simplifying trait in accordance to [PullupFrom].

PushdownInto
Syntax simplifying trait in accordance to [PushdownFrom].

ServiceCompliance
Trait for implementing a certain service on the state machine.

StackStorageCompliance
Defines stack behaviour for a certain storage object.

TransitionInto
Syntax simplifying trait in accordance to [TransitionFrom].

Functions

pack_transaction
Transform a transaction into the wrapping variant.

unpack_transaction
Unpack a wrapped transaction into an owned value.

Type Definitions

EntityId
Type that's generally used to identify and order Entity objects.