Module game_system::prelude::error::custom_type

Error types which represent one specific kind of error (failure). These errors are reported within functions that manipulate properties of the state machine, but will be wrapped into MachineError eventually.

Structs

IDCollisionError
Error thrown when the provided object's ID collides with an already known ID.

InvalidEntityMutUnwrap
Code failed to get a mutable reference to an Entity.

MissingCardError
Specific error thrown when the requested card-id is not known.

MissingEntityError
Specific error thrown when the requested entity-id is not known.

MissingPropertyError
Specific error thrown when the requested property is not known.

MissingPrototypeError
Specific error thrown when the requested entity-id is not known.

OverflowError
Specific error thrown to indicate the system cannot execute the request under constrained circumstances.

RuntimeConstraintError
Type used for indicating failure to meet specified constraints.

StackPopError
Code failed to push a new item onto the chosen stack.

StackPushError
Code failed to push a new item onto the chosen stack.

TransactionUnpackError
Code failed to push a new item onto the chosen stack.

Enums

TriggerFail
Enumeration of failure cases working with Triggers.