Trait game_system::prelude::function::IndexedStorageCompliance

pub trait IndexedStorageCompliance {
    type Item: Identifiable;
    fn get(
        &self, 
        identifier: <Self::Item as Identifiable>::ID
    ) -> Option<&Self::Item>;

    fn get_mut(
        &mut self, 
        identifier: <Self::Item as Identifiable>::ID
    ) -> Option<&mut Self::Item>;
}

Defines indexed behaviour for a certain storage object.

Associated Types

type Item: Identifiable

The type of items found within the implementing storage.

Required Methods

fn get(
&self,
identifier: <Self::Item as Identifiable>::ID
) -> Option<&Self::Item>

Retrieves a reference to the requested item matching the provided identifier.

fn get_mut(
&mut self,
identifier: <Self::Item as Identifiable>::ID
) -> Option<&mut Self::Item>

Retrieves a reference to the requested item matching the provided identifier.