Trait game_system::prelude::FrontendErrorExt

pub trait FrontendErrorExt {
    fn infuse<M, I>(self, context: ErrorKind, machine: I) -> MachineError
    where
        I: FnOnce() -> M,
        M: StateContainer + Debug + Send + 'static;

    fn infuse_with<F, M>(self, msg: F, machine: M) -> MachineError
    where
        F: Into<Cow<'static, str>>,
        M: StateContainer + Debug + Send + 'static;
}

Trait facilitating error creation with a snapshot of the state machine attached.

Required Methods

fn infuse<M, I>(self, context: ErrorKind, machine: I) -> MachineError where
I: FnOnce() -> M,
M: StateContainer + Debug + Send + 'static,

Builds a MachineError from some error or empty option.

Constraints

The error in question MUST implement [Fail]!

Parameters

context ErrorKind - is ment to categorize different errors. Make sure the value you choose is semantically correct because that's all the communicated information to the end user.

machine [´StateContainer] - is ment to store (effectively through [Clone`]) a snapshot of the state machine onto the heap. The stored state machine will be an exact copy of the real one at the moment of failure.

fn infuse_with<F, M>(self, msg: F, machine: M) -> MachineError where
F: Into<Cow<'static, str>>,
M: StateContainer + Debug + Send + 'static,

Builds a MachineError from some error or empty option. The ErrorKind invariant will be ErrorKind::Custom which will contain the string type passed into this method.

Parameters

custom [Cow] - represents an owned string or a string reference.

machine [´StateContainer] - is ment to store (effectively through [Clone`]) a snapshot of the state machine onto the heap. The stored state machine will be an exact copy of the real one at the moment of failure.