Trait game_system::prelude::Identifiable

pub trait Identifiable {
    type ID: Copy;
    fn id(&self) -> Self::ID;
}

Trait enforcing implementing objects to expose their identifier. The uniqueness of this identifier depends on the implementing object itself.

Associated Types

type ID: Copy

The type of identifier used to pass between functions during state machine execution. This type MUST be Copy because storing this identifier is the idiomatic way of passing "references" around.

Required Methods

fn id(&self) -> Self::ID

Returns the identifier of the implementing object.