Module game_system::re_export

Exported types from medici_core.

This crate declares macros which make direct use of the exported types so we make them available through this crate to keep the amount of dependancies small. This effectively removes the dependancy on medici_core for any downstream crate.

Re-exports

pub use medici_core::ct;

Modules

function
Contains the core functionality items for our system.

marker
Primitive traits which can be used as constraints by the core components.

service
Types that provide functionality to your application. These types can be added AD-HOC to your (state) machine declaration. Other components of this framework also need some of these services to be implemented before use.

storage
Types with each different functionality to store data.

Traits

PullupFrom
Trait defining the contract for two-way transitions of the state machine. The Transaction object of the outgoing state is stored for re-use. This kind of transition will write state change history into the compile-time stack ([CTStack]).

PushdownFrom
Trait defining the contract for two-way transitions of the state machine. The transition this trait defines is meant as a temporary one where the control flow will result in the state machine eventually performs a PullupFrom transition. Effectively reversing the Pushdown. The Transaction object of the outgoing state is stored for re-use. This kind of transition will write state change history into the compile-time stack ([CTStack]).

TransitionFrom
Trait defining the contract for one-way transitions of the state machine. This kind of transition preserves the state change history.