Trait game_system::re_export::function::ServiceCompliance

pub trait ServiceCompliance<S> where
    S: Service, {
    fn get(&self) -> &S;

    fn get_mut(&mut self) -> &mut S;
}

Trait for implementing a certain service on the state machine.

Because of this design exactly one object of each service type can be hooked onto the same state machine.

Required Methods

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

fn get(&self) -> &S

Retrieves an immutable reference to service S.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

fn get_mut(&mut self) -> &mut S

Retrieves a mutable reference to service S.