Trait game_system::re_export::function::State

pub trait State {
    type Transaction: Transaction;
}

Trait generalizing over any state that's present in the state machine.

Associated Types

type Transaction: Transaction

Type of structure which must be provided when transitioning into the state represented by the enclosing type.