Module game_system::re_export::marker

Primitive traits which can be used as constraints by the core components.

Marker Traits are usefull because the can be used as generic bounds. This allows for decoupling hidden code from developer created code. Correct understanding of what each trait encompasses is necessary!

Traits

Actionable
(State) Types which represent actions a user want to perform that activate the machine to work out its effects.

ProtoEnumerator
Types which enumerate all known [Prototype]s.

Prototype
Types which attribute functionality to [Entity]s within the machine.

Service
Types which attribute functionality to state machines.

Timing
Types which reflect the timing when a [Triggerable] should be executed.

TimingEnumerator
Types which enumerate all possible timings at which moment a trigger can be executed.

TopLevel
(State) Types which are directly contained by the state machine.

Transaction
Types used to transition between state machine States.

TransactionContainer
Types which generalize multiple transactions into 1 Sized structure so the transactions themselves can be safely stored in memory.

TriggerEnumerator
Types which enumerate all possible triggers which the machine facilitates reacting to.

Triggerable
(State) Types which reflect an event after which functionality awaiting these events will be executed.

Waitable
(State) Types which represent a condition for when the state machine itself should resume execution.