Struct game_system::re_export::service::trigger::TriggerService

pub struct TriggerService<ETM, ETR> where
    ETM: TimingEnumerator + PartialEq<ETM> + Copy,
    ETR: TriggerEnumerator + PartialEq<ETR> + Copy,  { /* fields omitted */ }

Structure wrapping and containing all [Trigger]s registered on the state machine.

Safety

This service erases the exact state-machine type by only keeping Timing and Trigger information. DO NOT execute callbacks built from a specific machine with another machine!

TODO: Think about enforcing this check at compile time.

Note: This could be enforced by letting the StateContainer return a unique number for each created object. This number could be used as constraint when reconstructing [TriggerWrapper].

Note: The trigger methods could be written as if they respond to a certain [Timing] and [Trigger] where the exact machine format is abstracted. Additionally, machine constraints can be defined on these methods to statically enforce services. This requires a lot of boilerplate code but macro's might be able to solve this?

eg:

#struct Post;
#impl Timing for Post {}

#struct EndTurn;
#impl Trigger for EndTurn {}

fn print_trigger<M>(machine: M) -> Result<M, Error>
where
    M: StateContainer,
    M::State: TriggerState<Timing=Post, Trigger=EndTurn>,
{
    unimplemented!()
}

Methods

impl<ETM, ETR> TriggerService<ETM, ETR> where
ETM: TimingEnumerator + PartialEq<ETM> + Copy,
ETR: TriggerEnumerator + PartialEq<ETR> + Copy,

pub fn new() -> TriggerService<ETM, ETR>

Creates a new object of this service.

pub fn add_trigger<M>(&mut self, cb: fn(M) -> Result<M, MachineError>) where
M: StateContainer,
<M as StateContainer>::State: TriggerState,
<<M as StateContainer>::State as TriggerState>::Timing: Timing,
<<M as StateContainer>::State as TriggerState>::Timing: IntoEnum<ETM>,
<<M as StateContainer>::State as TriggerState>::Trigger: Triggerable,
<<M as StateContainer>::State as TriggerState>::Trigger: IntoEnum<ETR>,

Add a new trigger to the store.

pub fn retrieve_all_triggers(
&self
) -> impl Iterator<Item = &UnsafeTrigger<ETM, ETR>>

Returns an iterator over all stored triggers.

pub fn retrieve_triggers<M>(
&self,
&M
) -> impl Iterator<Item = &UnsafeTrigger<ETM, ETR>> where
M: StateContainer,
<M as StateContainer>::State: TriggerState,
<<M as StateContainer>::State as TriggerState>::Timing: Timing,
<<M as StateContainer>::State as TriggerState>::Timing: IntoEnum<ETM>,
<<M as StateContainer>::State as TriggerState>::Trigger: Triggerable,
<<M as StateContainer>::State as TriggerState>::Trigger: IntoEnum<ETR>,

Retrieve all triggers matching the provided machine.

Borrow-check

This method might cause issues with the borrow checker because [Service] is designed to be contained by a state machine. By accessing this specific service we place an immutable borrow onto that machine, which is also passed as parameter into this method. In general this additional immutable borrow should not matter, but it does when there is already a mutable borrow of the machine. Returning [UnsafeTrigger] references will limit accessibility into the machine and this service.

The latter practically forces framework users to instantly make a copy of each returned [UnsafeTrigger] reference. Ultimately we leave the choice of usage up to the framework user. The reason being that we want additional functional operations to be as lightweight as possible.