Struct game_system::re_export::storage::trigger::UnsafeTrigger

pub struct UnsafeTrigger<ETM, ETR> where
    ETM: TimingEnumerator + Copy,
    ETR: TriggerEnumerator + Copy,  {
    pub timing: ETM,
    pub trigger: ETR,
    pub func_pointer: *const (),
    // some fields omitted
}

Structure serializng/generalizing a trigger.

The state machine will use this abstraction to store a specific method which has to be executed when the trigger conditions are valid. This type also makes triggers portable so they can be freely moved and copied through parts of the machine.

Fields
timing: ETM

(Sized) Timing value belonging to the callback, see func_pointer.

trigger: ETR

(Sized) Trigger value belonging to the callback, see func_pointer.

func_pointer: *const ()

The callback pointer which must be transmuted and executed when the conditions of the running state machine match the ones contained within this structure.