Trait game_system::re_export::PullupFrom

pub trait PullupFrom<T, CTS, TTC>: StateContainer<TransitionRecord = <CTS as CTStack>::Tail, TransitionRecord = <CTS as CTStack>::Tail> + 'static + StateContainer + 'static where
    CTS: CTStack + 'static,
    T: StateContainer<TransitionRecord = CTS> + 'static,
    TTC: TransactionContainer + 'static,
    Self::State: State,
    Self::State: 'static,
    <Self::State as State>::Transaction: Transaction,
    <Self::State as State>::Transaction: TryFrom<TTC>,
    <Self::State as State>::Transaction: 'static, {
    fn pullup_from(T) -> Result<Self, MachineError>;
}

Trait defining the contract for two-way transitions of the state machine. The Transaction object of the outgoing state is stored for re-use. This kind of transition will write state change history into the compile-time stack ([CTStack]).

A state machine is said to pullup from B into A when the current state is B, a [Transaction] object for state A is provided and the following transition is valid [A <- B].

Note

PushdownFrom is designed to be used together with PullupFrom because both operations store and reload the [Transaction] object respectively.

See also

PushdownFrom

Required Methods

fn pullup_from(T) -> Result<Self, MachineError>

Transition from the provided state into the implementing state.

Errors

Invalid transitions are blocked according to the state history log. There is also a runtime check which tests the type of the required transaction object against the last one stored (presumably by PushdownFrom) on a stack owned by the machine. If a [Transaction] type is popped that's not compatible with the incoming state an error is thrown.