Trait game_system::re_export::TransitionFrom

pub trait TransitionFrom<T, CTS>: StateContainer<TransitionRecord = CTS, TransitionRecord = CTS> + 'static + StateContainer + 'static where
    CTS: CTStack + 'static,
    T: StateContainer<TransitionRecord = CTS> + 'static,
    Self::State: State,
    Self::State: 'static,
    <Self::State as State>::Transaction: Transaction,
    <Self::State as State>::Transaction: 'static, {
    fn transition_from(T, <Self::State as State>::Transaction) -> Self;
}

Trait defining the contract for one-way transitions of the state machine. This kind of transition preserves the state change history.

A state machine is said to transition from A into B when the current state is A, a [Transaction] object for state B is provided and the following transition is valid [A -> B].

Required Methods

fn transition_from(T, <Self::State as State>::Transaction) -> Self

Transition from the provided state into the implementing state.