Struct game_system::state_machine::config::SetupConfig

pub struct SetupConfig {
    pub player_names: [Option<String>; 5],
    pub max_entities: usize,
}

Structure used for setting up a new [Machine].

Fields
player_names: [Option<String>; 5]

Name for each player.

This array is also used to calculate the amount of players to initialise. Make entries None to skip generation of a new player entity.

Note

PlayerID is a 1-indexed ordinal number. The first player defined as [Some(String)] will receive PlayerID 1.

max_entities: usize

Maximum amount of entities to be stored inside this machine.