Struct game_system::state_machine::machine::Machine

pub struct Machine<X, CTS> where
    X: TopLevel + State + Send,
    CTS: CTStack + Send,  {
    pub transaction: X::Transaction,
    pub triggers: TriggerService<TimingItem, TriggerItem>,
    pub entities: EntityService<Entity>,
    pub transactions: TransactionStorage<TransactionItem>,
    // some fields omitted
}

The state machine.

The developer is encouraged to alter this contents of this structure to fit his use case. Each state machine MUST have a state, history and transaction field AT MINIMUM.

Safety

Fields state and history are not accessible because altering them would cause the state transition and -history system to fall apart. Any code that has valid reason to alter these fields should be defined within this crate.

Fields
transaction: X::Transaction

Field to store the provided Transaction object as rquired by the current state.

triggers: TriggerService<TimingItem, TriggerItem>

Object for manipulating Triggers.

entities: EntityService<Entity>

Object for manipulating Entitys.

transactions: TransactionStorage<TransactionItem>

Storage object allowing [PushdownInto] and [PullupInto] to store the [Transaction] objects for each state to be re-used.

Methods

impl Machine<Wait<Start>, EmptyStack>

pub fn new(cfg: &SetupConfig) -> Result<Self, SetupError>

Creates a new state machine ready to be started.